+++ find_refs_arm32.cpp
@@ -1,7 +1,7 @@
 #include "find_refs.h"
 #include "binreader.h"
 
-void find_refs_arm32(size_t addr, std::vector<uint8_t> const & buf, std::function<void(size_t addr)> const & cb)
+void find_refs_arm32(uint64_t addr, std::vector<uint8_t> const & buf, std::function<void(uint64_t addr)> const & cb)
 {
 	if (addr % 4 != 0 || buf.size() % 4 != 0)
 		return;
